refactor: restructure configuration management with new types and default settings
